publish & update logic

+++ masterdata-service/service/src/main/java/de/uol/vpp/masterdata/service/services/ConsumerServiceImpl.java
@@ -2,25 +2,36 @@ package de.uol.vpp.masterdata.service.services;
 
 import de.uol.vpp.masterdata.domain.aggregates.HouseholdAggregate;
 import de.uol.vpp.masterdata.domain.entities.ConsumerEntity;
-import de.uol.vpp.masterdata.domain.repositories.ConsumerrRepositoryException;
+import de.uol.vpp.masterdata.domain.exceptions.ConsumerException;
+import de.uol.vpp.masterdata.domain.exceptions.HouseholdException;
+import de.uol.vpp.masterdata.domain.exceptions.VirtualPowerPlantException;
+import de.uol.vpp.masterdata.domain.repositories.ConsumerRepositoryException;
+import de.uol.vpp.masterdata.domain.repositories.HouseholdRepositoryException;
 import de.uol.vpp.masterdata.domain.repositories.IConsumerRepository;
 import de.uol.vpp.masterdata.domain.repositories.IHouseholdRepository;
 import de.uol.vpp.masterdata.domain.services.ConsumerServiceException;
 import de.uol.vpp.masterdata.domain.services.IConsumerService;
+import de.uol.vpp.masterdata.domain.utils.IPublishUtil;
+import de.uol.vpp.masterdata.domain.utils.PublishException;
 import de.uol.vpp.masterdata.domain.valueobjects.ConsumerIdVO;
+import de.uol.vpp.masterdata.domain.valueobjects.ConsumerStatusVO;
 import de.uol.vpp.masterdata.domain.valueobjects.HouseholdIdVO;
+import de.uol.vpp.masterdata.domain.valueobjects.VirtualPowerPlantIdVO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
+@Transactional(rollbackFor = ConsumerServiceException.class)
 @Service
 @RequiredArgsConstructor
 public class ConsumerServiceImpl implements IConsumerService {
 
     private final IConsumerRepository repository;
     private final IHouseholdRepository householdRepository;
+    private final IPublishUtil publishUtil;
 
     @Override
     public List<ConsumerEntity> getAllByHouseholdId(String householdBusinessKey) throws ConsumerServiceException {
@@ -33,7 +44,7 @@ public class ConsumerServiceImpl implements IConsumerService {
             throw new ConsumerServiceException(
                     String.format("There is no household %s to find any consumers", householdBusinessKey)
             );
-        } catch (ConsumerrRepositoryException e) {
+        } catch (HouseholdException | ConsumerRepositoryException | HouseholdRepositoryException e) {
             throw new ConsumerServiceException(e.getMessage(), e);
         }
     }
@@ -43,7 +54,7 @@ public class ConsumerServiceImpl implements IConsumerService {
         try {
             return repository.getById(new ConsumerIdVO(businessKey))
                     .orElseThrow(() -> new ConsumerServiceException(String.format("Can't find producer by id %s", businessKey)));
-        } catch (ConsumerrRepositoryException e) {
+        } catch (ConsumerRepositoryException | ConsumerException e) {
             throw new ConsumerServiceException(String.format("Can't find consumer by id %s", businessKey));
         }
 
@@ -52,6 +63,10 @@ public class ConsumerServiceImpl implements IConsumerService {
     @Override
     public void save(ConsumerEntity domainEntity, String householdBusinessKey) throws ConsumerServiceException {
         try {
+            if (repository.getById(domainEntity.getConsumerId()).isPresent()) {
+                throw new ConsumerServiceException(
+                        String.format("consumer with id %s already exists", domainEntity.getConsumerId().getId()));
+            }
             Optional<HouseholdAggregate> householdOptional = householdRepository.getById(
                     new HouseholdIdVO(householdBusinessKey)
             );
@@ -66,17 +81,49 @@ public class ConsumerServiceImpl implements IConsumerService {
                                 householdBusinessKey)
                 );
             }
-        } catch (ConsumerrRepositoryException e) {
+        } catch (ConsumerRepositoryException | HouseholdException | HouseholdRepositoryException e) {
             throw new ConsumerServiceException(e.getMessage(), e);
         }
     }
 
     @Override
-    public void delete(String businessKey) throws ConsumerServiceException {
+    public void delete(String businessKey, String vppBusinessKey) throws ConsumerServiceException {
         try {
-            repository.deleteById(new ConsumerIdVO(businessKey));
-        } catch (ConsumerrRepositoryException e) {
+            if (publishUtil.isEditable(new VirtualPowerPlantIdVO(vppBusinessKey), new ConsumerIdVO(businessKey))) {
+                repository.deleteById(new ConsumerIdVO(businessKey));
+            } else {
+                throw new ConsumerServiceException("failed to delete consumer. vpp has to be unpublished");
+            }
+        } catch (ConsumerRepositoryException | ConsumerException | VirtualPowerPlantException | PublishException e) {
             throw new ConsumerServiceException(e.getMessage(), e);
         }
     }
+
+    @Override
+    public void updateStatus(String businessKey, boolean isRunning, String vppBusinessKey) throws ConsumerServiceException {
+        try {
+            if (publishUtil.isEditable(new VirtualPowerPlantIdVO(vppBusinessKey), new ConsumerIdVO(vppBusinessKey))) {
+                repository.updateStatus(new ConsumerIdVO(businessKey),
+                        new ConsumerStatusVO(isRunning));
+            } else {
+                throw new ConsumerServiceException("failed to update consumer status. vpp has to be unpublished");
+            }
+
+        } catch (ConsumerRepositoryException | ConsumerException | VirtualPowerPlantException | PublishException e) {
+            throw new ConsumerServiceException(e.getMessage(), e);
+        }
+    }
+
+    @Override
+    public void update(String businessKey, ConsumerEntity domainEntity, String vppBusinessKey) throws ConsumerServiceException {
+        try {
+            if (publishUtil.isEditable(new VirtualPowerPlantIdVO(vppBusinessKey), new ConsumerIdVO(businessKey))) {
+                repository.update(new ConsumerIdVO(businessKey), domainEntity);
+            }
+        } catch (PublishException | VirtualPowerPlantException | ConsumerException | ConsumerRepositoryException e) {
+            throw new ConsumerServiceException(e.getMessage(), e);
+        }
+    }
+
+
 }
