+++ masterdata-service/service/src/main/java/de/uol/vpp/masterdata/service/services/DecentralizedPowerPlantServiceImpl.java
@@ -2,26 +2,33 @@ package de.uol.vpp.masterdata.service.services;
 
 import de.uol.vpp.masterdata.domain.aggregates.DecentralizedPowerPlantAggregate;
 import de.uol.vpp.masterdata.domain.aggregates.VirtualPowerPlantAggregate;
+import de.uol.vpp.masterdata.domain.exceptions.DecentralizedPowerPlantException;
+import de.uol.vpp.masterdata.domain.exceptions.VirtualPowerPlantException;
 import de.uol.vpp.masterdata.domain.repositories.DecentralizedPowerPlantRepositoryException;
 import de.uol.vpp.masterdata.domain.repositories.IDecentralizedPowerPlantRepository;
 import de.uol.vpp.masterdata.domain.repositories.IVirtualPowerPlantRepository;
 import de.uol.vpp.masterdata.domain.repositories.VirtualPowerPlantRepositoryException;
 import de.uol.vpp.masterdata.domain.services.DecentralizedPowerPlantServiceException;
 import de.uol.vpp.masterdata.domain.services.IDecentralizedPowerPlantService;
+import de.uol.vpp.masterdata.domain.utils.IPublishUtil;
+import de.uol.vpp.masterdata.domain.utils.PublishException;
 import de.uol.vpp.masterdata.domain.valueobjects.DecentralizedPowerPlantIdVO;
 import de.uol.vpp.masterdata.domain.valueobjects.VirtualPowerPlantIdVO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
+@Transactional(rollbackFor = DecentralizedPowerPlantServiceException.class)
 @RequiredArgsConstructor
 @Service
 public class DecentralizedPowerPlantServiceImpl implements IDecentralizedPowerPlantService {
 
     private final IDecentralizedPowerPlantRepository repository;
     private final IVirtualPowerPlantRepository virtualPowerPlantRepository;
+    private final IPublishUtil publishUtil;
 
     @Override
     public List<DecentralizedPowerPlantAggregate> getAllByVppId(String vppBusinessKey) throws DecentralizedPowerPlantServiceException {
@@ -34,25 +41,35 @@ public class DecentralizedPowerPlantServiceImpl implements IDecentralizedPowerPl
             throw new DecentralizedPowerPlantServiceException(
                     String.format("There is no vpp %s to find any dpp's", vppBusinessKey)
             );
-        } catch (VirtualPowerPlantRepositoryException e) {
+        } catch (VirtualPowerPlantRepositoryException | VirtualPowerPlantException | DecentralizedPowerPlantRepositoryException e) {
             throw new DecentralizedPowerPlantServiceException(e.getMessage(), e);
         }
-
     }
 
     @Override
     public DecentralizedPowerPlantAggregate get(String businessKey) throws DecentralizedPowerPlantServiceException {
-        return repository.getById(new DecentralizedPowerPlantIdVO(businessKey))
-                .orElseThrow(() -> new DecentralizedPowerPlantServiceException(String.format("Can't find DPP by id %s", businessKey)));
+        try {
+            Optional<DecentralizedPowerPlantAggregate> dpp = repository.getById(new DecentralizedPowerPlantIdVO(businessKey));
+            if (dpp.isPresent()) {
+                return dpp.get();
+            } else {
+                throw new DecentralizedPowerPlantServiceException(String.format("Can't find DPP by id %s", businessKey));
+            }
+        } catch (DecentralizedPowerPlantRepositoryException | DecentralizedPowerPlantException e) {
+            throw new DecentralizedPowerPlantServiceException(e.getMessage(), e);
+        }
     }
 
     @Override
     public void save(DecentralizedPowerPlantAggregate domainEntity, String virtualPowerPlantBusinessKey) throws DecentralizedPowerPlantServiceException {
         try {
+            if (repository.getById(domainEntity.getDecentralizedPowerPlantId()).isPresent()) {
+                throw new DecentralizedPowerPlantServiceException(
+                        String.format("dpp with id %s already exists", domainEntity.getDecentralizedPowerPlantId().getId()));
+            }
             Optional<VirtualPowerPlantAggregate> virtualPowerPlantOptional = virtualPowerPlantRepository.getById(
                     new VirtualPowerPlantIdVO(virtualPowerPlantBusinessKey)
             );
-
             if (virtualPowerPlantOptional.isPresent()) {
                 VirtualPowerPlantAggregate virtualPowerPlant = virtualPowerPlantOptional.get();
                 repository.save(domainEntity);
@@ -63,17 +80,35 @@ public class DecentralizedPowerPlantServiceImpl implements IDecentralizedPowerPl
                                 virtualPowerPlantBusinessKey)
                 );
             }
-        } catch (DecentralizedPowerPlantRepositoryException | VirtualPowerPlantRepositoryException e) {
+        } catch (DecentralizedPowerPlantRepositoryException | VirtualPowerPlantRepositoryException | VirtualPowerPlantException e) {
             throw new DecentralizedPowerPlantServiceException(e.getMessage(), e);
         }
     }
 
     @Override
-    public void delete(String businessKey) throws DecentralizedPowerPlantServiceException {
+    public void delete(String businessKey, String vppBusinessKey) throws DecentralizedPowerPlantServiceException {
         try {
-            repository.deleteById(new DecentralizedPowerPlantIdVO(businessKey));
-        } catch (DecentralizedPowerPlantRepositoryException e) {
+            if (publishUtil.isEditable(new VirtualPowerPlantIdVO(vppBusinessKey),
+                    new DecentralizedPowerPlantIdVO(businessKey))) {
+                repository.deleteById(new DecentralizedPowerPlantIdVO(businessKey));
+            } else {
+                throw new DecentralizedPowerPlantServiceException("deleting dpp failed. vpp has to be unpublished");
+            }
+        } catch (DecentralizedPowerPlantRepositoryException | DecentralizedPowerPlantException | VirtualPowerPlantException | PublishException e) {
             throw new DecentralizedPowerPlantServiceException(e.getMessage(), e);
         }
     }
+
+    @Override
+    public void update(String businessKey, DecentralizedPowerPlantAggregate domainEntity, String vppBusinessKey) throws DecentralizedPowerPlantServiceException {
+        try {
+            if (publishUtil.isEditable(new VirtualPowerPlantIdVO(vppBusinessKey), new DecentralizedPowerPlantIdVO(businessKey))) {
+                repository.update(new DecentralizedPowerPlantIdVO(businessKey), domainEntity);
+            }
+        } catch (PublishException | VirtualPowerPlantException | DecentralizedPowerPlantException | DecentralizedPowerPlantRepositoryException e) {
+            throw new DecentralizedPowerPlantServiceException(e.getMessage(), e);
+        }
+    }
+
+
 }
