+++ masterdata-service/service/src/main/java/de/uol/vpp/masterdata/service/services/HouseholdServiceImpl.java
@@ -2,26 +2,33 @@ package de.uol.vpp.masterdata.service.services;
 
 import de.uol.vpp.masterdata.domain.aggregates.HouseholdAggregate;
 import de.uol.vpp.masterdata.domain.aggregates.VirtualPowerPlantAggregate;
+import de.uol.vpp.masterdata.domain.exceptions.HouseholdException;
+import de.uol.vpp.masterdata.domain.exceptions.VirtualPowerPlantException;
 import de.uol.vpp.masterdata.domain.repositories.HouseholdRepositoryException;
 import de.uol.vpp.masterdata.domain.repositories.IHouseholdRepository;
 import de.uol.vpp.masterdata.domain.repositories.IVirtualPowerPlantRepository;
 import de.uol.vpp.masterdata.domain.repositories.VirtualPowerPlantRepositoryException;
 import de.uol.vpp.masterdata.domain.services.HouseholdServiceException;
 import de.uol.vpp.masterdata.domain.services.IHouseholdService;
+import de.uol.vpp.masterdata.domain.utils.IPublishUtil;
+import de.uol.vpp.masterdata.domain.utils.PublishException;
 import de.uol.vpp.masterdata.domain.valueobjects.HouseholdIdVO;
 import de.uol.vpp.masterdata.domain.valueobjects.VirtualPowerPlantIdVO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
+@Transactional(rollbackFor = HouseholdServiceException.class)
 @RequiredArgsConstructor
 @Service
 public class HouseholdServiceImpl implements IHouseholdService {
 
     private final IHouseholdRepository repository;
     private final IVirtualPowerPlantRepository virtualPowerPlantRepository;
+    private final IPublishUtil publishUtil;
 
     @Override
     public List<HouseholdAggregate> getAllByVppId(String vppBusinessKey) throws HouseholdServiceException {
@@ -32,9 +39,9 @@ public class HouseholdServiceImpl implements IHouseholdService {
                 return repository.getAllByVirtualPowerPlant(virtualPowerPlantOptional.get());
             }
             throw new HouseholdServiceException(
-                    String.format("There is no vpp %s to find any dpp's", vppBusinessKey)
+                    String.format("There is no vpp %s to find any households", vppBusinessKey)
             );
-        } catch (HouseholdRepositoryException | VirtualPowerPlantRepositoryException e) {
+        } catch (HouseholdRepositoryException | VirtualPowerPlantRepositoryException | VirtualPowerPlantException e) {
             throw new HouseholdServiceException(e.getMessage(), e);
         }
 
@@ -42,14 +49,22 @@ public class HouseholdServiceImpl implements IHouseholdService {
 
     @Override
     public HouseholdAggregate get(String businessKey) throws HouseholdServiceException {
-        return repository.getById(new HouseholdIdVO(businessKey))
-                .orElseThrow(() -> new HouseholdServiceException(String.format("Can't find household by id %s", businessKey)));
+        try {
+            return repository.getById(new HouseholdIdVO(businessKey))
+                    .orElseThrow(() -> new HouseholdServiceException(String.format("Can't find household by id %s", businessKey)));
+        } catch (HouseholdException | HouseholdRepositoryException e) {
+            throw new HouseholdServiceException(e.getMessage(), e);
+        }
 
     }
 
     @Override
     public void save(HouseholdAggregate domainEntity, String virtualPowerPlantBusinessKey) throws HouseholdServiceException {
         try {
+            if (repository.getById(domainEntity.getHouseholdId()).isPresent()) {
+                throw new HouseholdServiceException(
+                        String.format("household with id %s already exists", domainEntity.getHouseholdId().getId()));
+            }
             Optional<VirtualPowerPlantAggregate> virtualPowerPlantOptional = virtualPowerPlantRepository.getById(
                     new VirtualPowerPlantIdVO(virtualPowerPlantBusinessKey)
             );
@@ -60,20 +75,36 @@ public class HouseholdServiceImpl implements IHouseholdService {
                 repository.assign(domainEntity, virtualPowerPlant);
             } else {
                 throw new HouseholdServiceException(
-                        String.format("Failed to assign dpp %s to vpp %s", domainEntity.getHouseholdId().getId(),
+                        String.format("Failed to assign household %s to vpp %s", domainEntity.getHouseholdId().getId(),
                                 virtualPowerPlantBusinessKey)
                 );
             }
-        } catch (HouseholdRepositoryException | VirtualPowerPlantRepositoryException e) {
+        } catch (HouseholdRepositoryException | VirtualPowerPlantRepositoryException | VirtualPowerPlantException e) {
             throw new HouseholdServiceException(e.getMessage(), e);
         }
     }
 
     @Override
-    public void delete(String businessKey) throws HouseholdServiceException {
+    public void delete(String businessKey, String vppBusinessKey) throws HouseholdServiceException {
         try {
-            repository.deleteById(new HouseholdIdVO(businessKey));
-        } catch (HouseholdRepositoryException e) {
+            if (publishUtil.isEditable(new VirtualPowerPlantIdVO(vppBusinessKey), new HouseholdIdVO(businessKey))) {
+                repository.deleteById(new HouseholdIdVO(businessKey));
+            } else {
+                throw new HouseholdServiceException("failed to delete household. vpp has to be unpublished.");
+            }
+
+        } catch (HouseholdRepositoryException | HouseholdException | VirtualPowerPlantException | PublishException e) {
+            throw new HouseholdServiceException(e.getMessage(), e);
+        }
+    }
+
+    @Override
+    public void update(String businessKey, HouseholdAggregate domainEntity, String vppBusinessKey) throws HouseholdServiceException {
+        try {
+            if (publishUtil.isEditable(new VirtualPowerPlantIdVO(vppBusinessKey), new HouseholdIdVO(businessKey))) {
+                repository.update(new HouseholdIdVO(businessKey), domainEntity);
+            }
+        } catch (PublishException | VirtualPowerPlantException | HouseholdException | HouseholdRepositoryException e) {
             throw new HouseholdServiceException(e.getMessage(), e);
         }
     }
