+++ masterdata-service/service/src/main/java/de/uol/vpp/masterdata/service/services/ProducerServiceImpl.java
@@ -3,21 +3,24 @@ package de.uol.vpp.masterdata.service.services;
 import de.uol.vpp.masterdata.domain.aggregates.DecentralizedPowerPlantAggregate;
 import de.uol.vpp.masterdata.domain.aggregates.HouseholdAggregate;
 import de.uol.vpp.masterdata.domain.entities.ProducerEntity;
-import de.uol.vpp.masterdata.domain.repositories.IDecentralizedPowerPlantRepository;
-import de.uol.vpp.masterdata.domain.repositories.IHouseholdRepository;
-import de.uol.vpp.masterdata.domain.repositories.IProducerRepository;
-import de.uol.vpp.masterdata.domain.repositories.ProducerRepositoryException;
+import de.uol.vpp.masterdata.domain.exceptions.DecentralizedPowerPlantException;
+import de.uol.vpp.masterdata.domain.exceptions.HouseholdException;
+import de.uol.vpp.masterdata.domain.exceptions.ProducerException;
+import de.uol.vpp.masterdata.domain.exceptions.VirtualPowerPlantException;
+import de.uol.vpp.masterdata.domain.repositories.*;
 import de.uol.vpp.masterdata.domain.services.IProducerService;
 import de.uol.vpp.masterdata.domain.services.ProducerServiceException;
-import de.uol.vpp.masterdata.domain.valueobjects.DecentralizedPowerPlantIdVO;
-import de.uol.vpp.masterdata.domain.valueobjects.HouseholdIdVO;
-import de.uol.vpp.masterdata.domain.valueobjects.ProducerIdVO;
+import de.uol.vpp.masterdata.domain.utils.IPublishUtil;
+import de.uol.vpp.masterdata.domain.utils.PublishException;
+import de.uol.vpp.masterdata.domain.valueobjects.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
+@Transactional(rollbackFor = ProducerServiceException.class)
 @Service
 @RequiredArgsConstructor
 public class ProducerServiceImpl implements IProducerService {
@@ -25,6 +28,7 @@ public class ProducerServiceImpl implements IProducerService {
     private final IProducerRepository repository;
     private final IDecentralizedPowerPlantRepository decentralizedPowerPlantRepository;
     private final IHouseholdRepository householdRepository;
+    private final IPublishUtil publishUtil;
 
     @Override
     public List<ProducerEntity> getAllByDecentralizedPowerPlantId(String dppBusinessKey) throws ProducerServiceException {
@@ -36,7 +40,7 @@ public class ProducerServiceImpl implements IProducerService {
             throw new ProducerServiceException(
                     String.format("There is no dpp %s to find any producers", dppBusinessKey)
             );
-        } catch (ProducerRepositoryException e) {
+        } catch (ProducerRepositoryException | DecentralizedPowerPlantException | DecentralizedPowerPlantRepositoryException e) {
             throw new ProducerServiceException(e.getMessage(), e);
         }
     }
@@ -52,7 +56,7 @@ public class ProducerServiceImpl implements IProducerService {
             throw new ProducerServiceException(
                     String.format("There is no household %s to find any producers", householdBusinessKey)
             );
-        } catch (ProducerRepositoryException e) {
+        } catch (ProducerRepositoryException | HouseholdException | HouseholdRepositoryException e) {
             throw new ProducerServiceException(e.getMessage(), e);
         }
     }
@@ -62,7 +66,7 @@ public class ProducerServiceImpl implements IProducerService {
         try {
             return repository.getById(new ProducerIdVO(businessKey))
                     .orElseThrow(() -> new ProducerServiceException(String.format("Can't find producer by id %s", businessKey)));
-        } catch (ProducerRepositoryException e) {
+        } catch (ProducerException | ProducerRepositoryException e) {
             throw new ProducerServiceException(String.format("Can't find producer by id %s", businessKey));
         }
 
@@ -71,6 +75,10 @@ public class ProducerServiceImpl implements IProducerService {
     @Override
     public void saveWithDecentralizedPowerPlant(ProducerEntity domainEntity, String dppBusinessKey) throws ProducerServiceException {
         try {
+            if (repository.getById(domainEntity.getProducerId()).isPresent()) {
+                throw new ProducerServiceException(
+                        String.format("producer with id %s already exists", domainEntity.getProducerId().getId()));
+            }
             Optional<DecentralizedPowerPlantAggregate> dppOptional = decentralizedPowerPlantRepository.getById(
                     new DecentralizedPowerPlantIdVO(dppBusinessKey)
             );
@@ -85,7 +93,7 @@ public class ProducerServiceImpl implements IProducerService {
                                 dppBusinessKey)
                 );
             }
-        } catch (ProducerRepositoryException e) {
+        } catch (ProducerRepositoryException | DecentralizedPowerPlantException | DecentralizedPowerPlantRepositoryException e) {
             throw new ProducerServiceException(e.getMessage(), e);
         }
     }
@@ -93,6 +101,10 @@ public class ProducerServiceImpl implements IProducerService {
     @Override
     public void saveWithHousehold(ProducerEntity domainEntity, String householdBusinessKey) throws ProducerServiceException {
         try {
+            if (repository.getById(domainEntity.getProducerId()).isPresent()) {
+                throw new ProducerServiceException(
+                        String.format("producer with id %s already exists", domainEntity.getProducerId().getId()));
+            }
             Optional<HouseholdAggregate> householdOptional = householdRepository.getById(
                     new HouseholdIdVO(householdBusinessKey)
             );
@@ -107,17 +119,50 @@ public class ProducerServiceImpl implements IProducerService {
                                 householdBusinessKey)
                 );
             }
-        } catch (ProducerRepositoryException e) {
+        } catch (ProducerRepositoryException | HouseholdException | HouseholdRepositoryException e) {
             throw new ProducerServiceException(e.getMessage(), e);
         }
     }
 
     @Override
-    public void delete(String businessKey) throws ProducerServiceException {
+    public void delete(String businessKey, String vppBusinessKey) throws ProducerServiceException {
         try {
-            repository.deleteById(new ProducerIdVO(businessKey));
-        } catch (ProducerRepositoryException e) {
+            if (publishUtil.isEditable(new VirtualPowerPlantIdVO(vppBusinessKey), new ProducerIdVO(businessKey))) {
+                repository.deleteById(new ProducerIdVO(businessKey));
+            } else {
+                throw new ProducerServiceException("failed to delete producer. vpp has to be unpublished");
+            }
+        } catch (ProducerRepositoryException | ProducerException | VirtualPowerPlantException | PublishException e) {
             throw new ProducerServiceException(e.getMessage(), e);
         }
     }
+
+    @Override
+    public void updateStatus(String businessKey, Double capacity, boolean running, String vppBusinessKey) throws ProducerServiceException {
+        try {
+            if (publishUtil.isEditable(new VirtualPowerPlantIdVO(vppBusinessKey), new ProducerIdVO(businessKey))) {
+                repository.updateStatus(
+                        new ProducerIdVO(businessKey),
+                        new ProducerStatusVO(running, capacity)
+                );
+            } else {
+                throw new ProducerServiceException("failed to delete producer. vpp has to be unpublished");
+            }
+
+        } catch (ProducerRepositoryException | ProducerException | VirtualPowerPlantException | PublishException e) {
+            throw new ProducerServiceException(e.getMessage(), e);
+        }
+    }
+
+    @Override
+    public void update(String businessKey, ProducerEntity domainEntity, String vppBusinessKey) throws ProducerServiceException {
+        try {
+            if (publishUtil.isEditable(new VirtualPowerPlantIdVO(vppBusinessKey), new ProducerIdVO(businessKey))) {
+                repository.update(new ProducerIdVO(businessKey), domainEntity);
+            }
+        } catch (PublishException | VirtualPowerPlantException | ProducerException | ProducerRepositoryException e) {
+            throw new ProducerServiceException(e.getMessage(), e);
+        }
+    }
+
 }
